requirements.txt
annotated-types==0.6.0
anyascii==0.3.2
beautifulsoup4==4.12.2
bs4==0.0.1
certifi==2023.7.22
cffi==1.16.0
chardet==5.0.0
charset-normalizer==3.3.1
Click==7.0
click-option-group==0.5.3
colorama==0.4.6
contractions==0.1.73
cryptography==41.0.5
cycler==0.12.1
decorator==5.1.1
Deprecated==1.2.14
duckdb==0.9.1
elementpath==4.1.5
exceptiongroup==1.1.3
falcon==3.1.1
fonttools==4.43.1
greenlet==3.0.1
idna==3.4
imbalanced-learn==0.8.1
importlib-metadata==6.8.0
inflect==7.0.0
iniconfig==2.0.0
isodate==0.6.1
joblib==1.3.2
jsonpath-python==1.0.6
jsonpointer==2.4
kiwisolver==1.4.5
Markdown==3.3.6
matplotlib==3.5.0
morph-kgc==2.3.1
nltk==3.6.6
numpy==1.22.0
packaging==23.2
pandas==1.3.4
Pillow==10.1.0
pluggy==1.3.0
pyahocorasick==2.0.0
pycparser==2.21
pydantic==2.4.2
pydantic_core==2.10.1
PyGithub==2.1.1
PyJWT==2.8.0
PyNaCl==1.5.0
pyoxigraph==0.3.20
pyparsing==3.1.1
pytest==7.4.3
python-dateutil==2.8.2
pytz==2023.3.post1
rdflib==7.0.0
rdflib-jsonld==0.6.2
regex==2023.10.3
requests==2.31.0
scikit-learn==1.0
scipy==1.11.3
setuptools-scm==8.0.4
six==1.16.0
somef==0.9.4
soupsieve==2.5
sql_metadata==2.9.0
SQLAlchemy==2.0.22
sqlparse==0.4.4
textblob==0.17.1
textsearch==0.0.24
threadpoolctl==3.2.0
tomli==2.0.1
tqdm==4.66.1
typing_extensions==4.8.0
urllib3==2.0.7
validators==0.18.2
wrapt==1.15.0
xgboost==1.5.0
XlsxWriter==3.1.9
zipp==3.17.0
